add Unit & Integration Testing flags
- instrument commands/ package
