+++ commands/actor_daemon_test.go
@@ -7,13 +7,15 @@ import (
 
 	"github.com/filecoin-project/go-filecoin/commands"
 	th "github.com/filecoin-project/go-filecoin/testhelpers"
+	tf "github.com/filecoin-project/go-filecoin/testhelpers/testflags"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestActorDaemon(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	t.Run("actor ls --enc json returns NDJSON containing all actors in the state tree", func(t *testing.T) {
 		require := require.New(t)
 		assert := assert.New(t)
