+++ commands/client_daemon_test.go
@@ -8,11 +8,13 @@ import (
 	"github.com/filecoin-project/go-filecoin/fixtures"
 	"github.com/filecoin-project/go-filecoin/protocol/storage"
 	th "github.com/filecoin-project/go-filecoin/testhelpers"
+	tf "github.com/filecoin-project/go-filecoin/testhelpers/testflags"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestListAsks(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	minerDaemon := makeTestDaemonWithMinerAndStart(t)
@@ -26,6 +28,8 @@ func TestListAsks(t *testing.T) {
 }
 
 func TestStorageDealsAfterRestart(t *testing.T) {
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 	minerDaemon := th.NewDaemon(t,
 		th.WithMiner(fixtures.TestMiners[0]),
@@ -67,7 +71,8 @@ func TestStorageDealsAfterRestart(t *testing.T) {
 }
 
 func TestDuplicateDeals(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	miner := th.NewDaemon(t,
@@ -103,7 +108,8 @@ func TestDuplicateDeals(t *testing.T) {
 }
 
 func TestDealWithSameDataAndDifferentMiners(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	miner1Addr := fixtures.TestMiners[0]
@@ -148,7 +154,8 @@ func TestDealWithSameDataAndDifferentMiners(t *testing.T) {
 }
 
 func TestVoucherPersistenceAndPayments(t *testing.T) {
-	t.Parallel()
+	tf.IntegrationTest(t)
+
 	assert := assert.New(t)
 
 	// DefaultAddress required here
