+++ commands/config_daemon_test.go
@@ -1,4 +1,4 @@
-package commands
+package commands_test
 
 import (
 	"encoding/json"
@@ -7,14 +7,16 @@ import (
 
 	"github.com/filecoin-project/go-filecoin/config"
 	th "github.com/filecoin-project/go-filecoin/testhelpers"
+	tf "github.com/filecoin-project/go-filecoin/testhelpers/testflags"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestConfigDaemon(t *testing.T) {
+	tf.IntegrationTest(t)
+
 	t.Run("config <key> prints config value", func(t *testing.T) {
-		t.Parallel()
 		assert := assert.New(t)
 		require := require.New(t)
 
@@ -35,7 +37,6 @@ func TestConfigDaemon(t *testing.T) {
 	})
 
 	t.Run("config <key> simple_value updates config", func(t *testing.T) {
-		t.Parallel()
 		assert := assert.New(t)
 
 		d := th.NewDaemon(t).Start()
@@ -57,7 +58,6 @@ func TestConfigDaemon(t *testing.T) {
 	})
 
 	t.Run("config <key> <val> updates config", func(t *testing.T) {
-		t.Parallel()
 		assert := assert.New(t)
 		require := require.New(t)
