+++ commands/from_address_test.go
@@ -6,19 +6,19 @@ import (
 	"github.com/ipfs/go-ipfs-cmdkit"
 
 	"github.com/filecoin-project/go-filecoin/address"
+	tf "github.com/filecoin-project/go-filecoin/testhelpers/testflags"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestOptionalAddr(t *testing.T) {
-	t.Parallel()
+	tf.UnitTest(t)
+
 	assert := assert.New(t)
 	require := require.New(t)
 
 	t.Run("when option is specified", func(t *testing.T) {
-		t.Parallel()
-
 		opts := make(cmdkit.OptMap)
 
 		specifiedAddr, err := address.NewActorAddress([]byte("a new test address"))
@@ -31,8 +31,6 @@ func TestOptionalAddr(t *testing.T) {
 	})
 
 	t.Run("when no option specified return empty", func(t *testing.T) {
-		t.Parallel()
-
 		opts := make(cmdkit.OptMap)
 
 		addr, err := optionalAddr(opts["from"])
